Created test case for Order model
